+++ objectiveai-rs/src/upstream/claude_agent_sdk/beta_raw_message_stream_event/beta_raw_content_block_start_event.rs
@@ -9,6 +9,6 @@ pub enum BetaRawContentBlockStartEventType {
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
 pub struct BetaRawContentBlockStartEvent {
     pub content_block: super::super::beta_content_block::BetaContentBlock,
-    pub index: f64,
+    pub index: i64,
     pub r#type: BetaRawContentBlockStartEventType,
 }
