+++ objectiveai-rs/src/upstream/claude_agent_sdk/beta_usage/beta_cache_creation.rs
@@ -2,6 +2,6 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
 pub struct BetaCacheCreation {
-    pub ephemeral_1h_input_tokens: f64,
-    pub ephemeral_5m_input_tokens: f64,
+    pub ephemeral_1h_input_tokens: i64,
+    pub ephemeral_5m_input_tokens: i64,
 }
