+++ objectiveai-rs/src/upstream/claude_agent_sdk/beta_usage/beta_message_iteration_usage.rs
@@ -9,9 +9,9 @@ pub enum BetaMessageIterationUsageType {
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
 pub struct BetaMessageIterationUsage {
     pub cache_creation: Option<super::BetaCacheCreation>,
-    pub cache_creation_input_tokens: f64,
-    pub cache_read_input_tokens: f64,
-    pub input_tokens: f64,
-    pub output_tokens: f64,
+    pub cache_creation_input_tokens: i64,
+    pub cache_read_input_tokens: i64,
+    pub input_tokens: i64,
+    pub output_tokens: i64,
     pub r#type: BetaMessageIterationUsageType,
 }
