+++ objectiveai-rs/src/upstream/claude_agent_sdk/beta_usage/beta_server_tool_usage.rs
@@ -2,6 +2,6 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
 pub struct BetaServerToolUsage {
-    pub web_fetch_requests: f64,
-    pub web_search_requests: f64,
+    pub web_fetch_requests: i64,
+    pub web_search_requests: i64,
 }
