+++ objectiveai-rs/src/upstream/claude_agent_sdk/beta_usage/beta_usage.rs
@@ -18,12 +18,12 @@ pub enum Speed {
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
 pub struct BetaUsage {
     pub cache_creation: Option<super::BetaCacheCreation>,
-    pub cache_creation_input_tokens: Option<f64>,
-    pub cache_read_input_tokens: Option<f64>,
+    pub cache_creation_input_tokens: Option<i64>,
+    pub cache_read_input_tokens: Option<i64>,
     pub inference_geo: Option<String>,
-    pub input_tokens: f64,
+    pub input_tokens: i64,
     pub iterations: Option<super::BetaIterationsUsage>,
-    pub output_tokens: f64,
+    pub output_tokens: i64,
     pub server_tool_use: Option<super::BetaServerToolUsage>,
     pub service_tier: Option<ServiceTier>,
     pub speed: Option<Speed>,
@@ -32,12 +32,12 @@ pub struct BetaUsage {
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
 pub struct NonNullableBetaUsage {
     pub cache_creation: super::BetaCacheCreation,
-    pub cache_creation_input_tokens: f64,
-    pub cache_read_input_tokens: f64,
+    pub cache_creation_input_tokens: i64,
+    pub cache_read_input_tokens: i64,
     pub inference_geo: String,
-    pub input_tokens: f64,
+    pub input_tokens: i64,
     pub iterations: super::BetaIterationsUsage,
-    pub output_tokens: f64,
+    pub output_tokens: i64,
     pub server_tool_use: super::BetaServerToolUsage,
     pub service_tier: ServiceTier,
     pub speed: Speed,
