+++ objectiveai-rs/src/upstream/claude_agent_sdk/content_block_param/text_citation_param.rs
@@ -9,10 +9,10 @@ pub enum CitationCharLocationParamType {
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
 pub struct CitationCharLocationParam {
     pub cited_text: String,
-    pub document_index: f64,
+    pub document_index: i64,
     pub document_title: Option<String>,
-    pub end_char_index: f64,
-    pub start_char_index: f64,
+    pub end_char_index: i64,
+    pub start_char_index: i64,
     pub r#type: CitationCharLocationParamType,
 }
 
@@ -25,10 +25,10 @@ pub enum CitationPageLocationParamType {
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
 pub struct CitationPageLocationParam {
     pub cited_text: String,
-    pub document_index: f64,
+    pub document_index: i64,
     pub document_title: Option<String>,
-    pub end_page_number: f64,
-    pub start_page_number: f64,
+    pub end_page_number: i64,
+    pub start_page_number: i64,
     pub r#type: CitationPageLocationParamType,
 }
 
@@ -41,10 +41,10 @@ pub enum CitationContentBlockLocationParamType {
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
 pub struct CitationContentBlockLocationParam {
     pub cited_text: String,
-    pub document_index: f64,
+    pub document_index: i64,
     pub document_title: Option<String>,
-    pub end_block_index: f64,
-    pub start_block_index: f64,
+    pub end_block_index: i64,
+    pub start_block_index: i64,
     pub r#type: CitationContentBlockLocationParamType,
 }
 
@@ -72,10 +72,10 @@ pub enum CitationSearchResultLocationParamType {
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
 pub struct CitationSearchResultLocationParam {
     pub cited_text: String,
-    pub end_block_index: f64,
-    pub search_result_index: f64,
+    pub end_block_index: i64,
+    pub search_result_index: i64,
     pub source: String,
-    pub start_block_index: f64,
+    pub start_block_index: i64,
     pub title: Option<String>,
     pub r#type: CitationSearchResultLocationParamType,
 }
